requirements.txt
# Core
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Web UI
streamlit>=1.30.0

# Document Parsing
PyMuPDF>=1.23.0
python-docx>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0

# ML / Embeddings / Reranking
sentence-transformers>=2.2.0
numpy>=1.24.0

# OCR — Image text extraction
pytesseract>=0.3.10
Pillow>=10.0.0

# Vector Database
opensearch-py>=2.4.0

# LLM
groq>=0.4.0
